requirements.txt
dash==2.14.1
plotly==5.17.0
pandas==2.1.1
requests==2.31.0
python-dotenv==1.0.0
dash-bootstrap-components==1.5.0
numpy==1.24.3
gunicorn==21.2.0
